version1/requirements.txt
flask
flask-cors
groq
fastembed
chromadb
python-dotenv
colorama
pyaudio
webrtcvad
numpy
google-generativeai
